# Caching of DataPointIterator instances