--- a/Directory.Build.props
+++ b/Directory.Build.props
@@ -15,7 +15,7 @@
 
     <PropertyGroup>
         <LangVersion>latest</LangVersion>
-        <StandardTfms>netcoreapp3.1</StandardTfms>
+        <StandardTfms>net5.0</StandardTfms>
     </PropertyGroup>
 
     <PropertyGroup>