--- a/perf/Directory.Build.props
+++ b/perf/Directory.Build.props
@@ -4,8 +4,8 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <StandardTfm>netcoreapp3.1</StandardTfm>
-        <LangVersion>9</LangVersion>
+        <StandardTfm>net5.0</StandardTfm>
+        <LangVersion>9.0</LangVersion>
         <Nullable>enable</Nullable>
     </PropertyGroup>