--- a/perf/gfoidl.DataCompression.Benchmarks/gfoidl.DataCompression.Benchmarks.csproj
+++ b/perf/gfoidl.DataCompression.Benchmarks/gfoidl.DataCompression.Benchmarks.csproj
@@ -1,16 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <OutputType>Exe</OutputType>
         <TargetFramework>$(StandardTfms)</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\..\source\gfoidl.DataCompression\gfoidl.DataCompression.csproj" />
     </ItemGroup>
 
 </Project>