--- a/perf/gfoidl.DataCompression.Benchmarks/gfoidl.DataCompression.Benchmarks.csproj
+++ b/perf/gfoidl.DataCompression.Benchmarks/gfoidl.DataCompression.Benchmarks.csproj
@@ -5,7 +5,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
+        <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
     </ItemGroup>
 
 </Project>