--- a/perf/gfoidl.DataCompression.ProfilingDemo/gfoidl.DataCompression.ProfilingDemo.csproj
+++ b/perf/gfoidl.DataCompression.ProfilingDemo/gfoidl.DataCompression.ProfilingDemo.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>$(StandardTfm)</TargetFramework>
+    </PropertyGroup>
+
+</Project>