--- a/source/gfoidl.DataCompression/Compression/DeadBandCompression/DeadBandCompressionIterator.cs
+++ b/source/gfoidl.DataCompression/Compression/DeadBandCompression/DeadBandCompressionIterator.cs
@@ -1,12 +1,13 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
 namespace gfoidl.DataCompression.Internal.DeadBand
 {
     internal abstract class DeadBandCompressionIterator : DataPointIterator
     {
-        protected readonly DeadBandCompression _deadBandCompression;
-        protected (double Min, double Max)     _bounding;
+        protected DeadBandCompression?     _deadBandCompression;
+        protected (double Min, double Max) _bounding;
         //---------------------------------------------------------------------
         protected DeadBandCompressionIterator(DeadBandCompression deadBandCompression)
             : base(deadBandCompression)
@@ -15,6 +16,8 @@ namespace gfoidl.DataCompression.Internal.DeadBand
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         protected void GetBounding(in DataPoint dataPoint)
         {
+            Debug.Assert(_deadBandCompression is not null);
+
             double y = dataPoint.Y;
 
             // Produces better code than updating _bounding directly
@@ -48,5 +51,13 @@ namespace gfoidl.DataCompression.Internal.DeadBand
         //---------------------------------------------------------------------
         protected internal override void Init(in DataPoint incoming, ref DataPoint snapShot)                   => this.UpdatePoints(incoming, ref snapShot);
         protected internal override void Init(int incomingIndex, in DataPoint incoming, ref int snapShotIndex) => throw new NotSupportedException();
+        //---------------------------------------------------------------------
+        protected override void ResetToInitialState()
+        {
+            base.ResetToInitialState();
+
+            _deadBandCompression = null;
+            _bounding            = default;
+        }
     }
 }