--- a/source/gfoidl.DataCompression/Compression/NoCompression/NoCompressionIterator.cs
+++ b/source/gfoidl.DataCompression/Compression/NoCompression/NoCompressionIterator.cs
@@ -1,9 +1,5 @@
 ﻿namespace gfoidl.DataCompression.Internal.NoCompression
 {
     internal abstract class NoCompressionIterator : DataPointIterator
-    {
-        protected NoCompressionIterator(Compression compression)
-            : base(compression)
-        { }
-    }
+    { }
 }