… .circleci/config.yml
… source/gfoidl.DataCompression/Compression/DeadBandCompression/IndexedIterator.cs
… source/gfoidl.DataCompression/Compression/NoCompression/AsyncEnumerableIterator.cs
… source/gfoidl.DataCompression/Compression/NoCompression/EnumerableIterator.cs
… source/gfoidl.DataCompression/Compression/NoCompression/NoCompression.cs
… source/gfoidl.DataCompression/Compression/SwingingDoorCompression/AsyncEnumerableIterator.cs
… source/gfoidl.DataCompression/Compression/SwingingDoorCompression/IndexedIterator.cs
… source/gfoidl.DataCompression/Compression/SwingingDoorCompression/SequentialEnumerableIterator.cs
… source/gfoidl.DataCompression/Compression/SwingingDoorCompression/SwingingDoorCompression.cs
… source/gfoidl.DataCompression/Compression/SwingingDoorCompression/SwingingDoorCompressionIterator.cs
… source/gfoidl.DataCompression/DataPointIndexedIterator.cs
… source/gfoidl.DataCompression/DataPointIterator.Async.cs
… source/gfoidl.DataCompression/DataPointIterator.Enumerable.cs
… source/gfoidl.DataCompression/EmptyDataPointIterator.cs
… source/gfoidl.DataCompression/InternalsVisibleTo.cs
… source/gfoidl.DataCompression/ThrowHelper.cs
… source/gfoidl.DataCompression/gfoidl.DataCompression.csproj
… tests/Directory.Build.props
… tests/gfoidl.DataCompression.Tests/Compression/DeadBandCompressionTests/IteratorCaching.cs
… tests/gfoidl.DataCompression.Tests/Compression/SwingingDoorCompressionTests/IteratorCaching.cs