## Description

Made the creation of the `DataPointIterator`s amortized allocation free.

Usage patterns like
```c#
DeadBandCompression deadBand         = new DeadBandCompression(0.1);
SwingingDoorCompression swingingDoor = new SwingingDoorCompression(0.1);

while (...)
{
    DataPointIterator deadBandCompressed     = deadBand.Process(rawDataPoints);
    DataPointIterator swingingDoorCompressed = swingingDoor.Process(deadBandCompressed);

    foreach (DataPoint item in swingingDoorCompressed)
    {
        compressed.Add(item);
    }
}
```
where the iterator is created in a loop itself, it's now without further allocations.

## Results

### Before

![grafik](https://user-images.githubusercontent.com/5755834/101844008-b1698280-3b4b-11eb-9d50-7a1a6ae3970e.png)

```
|             Method |    n |         Mean |         Error |       StdDev |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |----- |-------------:|--------------:|-------------:|---------:|------:|------:|----------:|
|    SingleIteration |    ? |     119.4 ns |      35.37 ns |      1.94 ns |   0.1018 |     - |     - |     320 B |
| MultipleIterations |  100 |  14,547.9 ns |   9,871.65 ns |    541.10 ns |  10.1929 |     - |     - |   32000 B |
| MultipleIterations | 1000 | 147,214.0 ns | 309,726.03 ns | 16,977.13 ns | 101.8066 |     - |     - |  320000 B |
```
[this benchmark](https://github.com/gfoidl/DataCompression/blob/fd84b257b5aaed967c7abe3db4d4771ca103f3bd/perf/gfoidl.DataCompression.Benchmarks/IteratorInstantiatonBenchmark.cs#L10)

A perfect O(n).

### After

![grafik](https://user-images.githubusercontent.com/5755834/101844055-dbbb4000-3b4b-11eb-8959-a7947fa27920.png)

```
|             Method |    n |         Mean |         Error |       StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |----- |-------------:|--------------:|-------------:|------:|------:|------:|----------:|
|    SingleIteration |    ? |     110.7 ns |      93.31 ns |      5.11 ns |     - |     - |     - |         - |
| MultipleIterations |  100 |  10,772.9 ns |   8,084.66 ns |    443.15 ns |     - |     - |     - |         - |
| MultipleIterations | 1000 | 129,554.4 ns | 536,653.22 ns | 29,415.77 ns |     - |     - |     - |         - |
```
[this benchmark](https://github.com/gfoidl/DataCompression/blob/fd84b257b5aaed967c7abe3db4d4771ca103f3bd/perf/gfoidl.DataCompression.Benchmarks/IteratorInstantiatonBenchmark.cs#L10)